refactor: Update API structure and introduce new evaluation models for enhanced data handling
